requirements.txt
fastapi[standard]
psycopg2
SQLAlchemy
authlib
python-dotenv
python-multipart
starlette-sessions
itsdangerous
uvicorn
alembic